+++ src/utils/mediaShape.js
@@ -1,3 +1,0 @@
-import { PropTypes } from 'react'
-
-export default PropTypes.objectOf(PropTypes.object)